+++ R/arkhe-package.R
@@ -39,4 +39,5 @@
 #' @importFrom methods callGeneric callNextMethod setGeneric setMethod
 #' .valueClassTest
 #' @importFrom stats quantile na.omit rmultinom sd xtabs
+#' @importFrom utils install.packages
 NULL