+++ R/assert.R
@@ -2,6 +2,49 @@
 #' @include predicates.R
 NULL
 
+#' Check the Availability of a Package
+#'
+#' @param x A [`character`] vector naming the packages to check.
+#' @details
+#'  `needs()` is designed for use inside other functions in your own package
+#'  to check for the availability of a suggested package.
+#'
+#'  If the required packages are not available and \R is running interactively,
+#'  the user will be asked to install the packages.
+#' @return Invisibly returns `NULL`.
+#' @family validation methods
+#' @author N. Frerebeau
+#' @export
+needs <- function(x) {
+  ok <- vapply(X = x, FUN = requireNamespace, FUN.VALUE = logical(1),
+               quietly = TRUE)
+
+  miss <- x[!ok]
+  n <- length(miss)
+
+  if (n != 0) {
+    msg <- ngettext(n, "Package %s is required.", "Packages %s are required.")
+    pkg <- paste0(sQuote(miss), collapse = ", ")
+    install <- "0"
+    if (interactive()) {
+      cat(
+        sprintf(msg, pkg),
+        sprintf("Do you want to install %s?", ngettext(n, "it", "them")),
+        "1. Yes",
+        "2. No",
+        sep = "\n"
+      )
+      install <- readline("Choice: ")
+    }
+    if (install == "1") {
+      utils::install.packages(miss)
+    } else {
+      stop(sprintf(msg, pkg), call. = FALSE)
+    }
+  }
+  invisible(NULL)
+}
+
 #' Validate a Condition
 #'
 #' @param expr An object to be evaluated.