+++ NEWS.md
@@ -1,4 +1,6 @@
 # arkhe 1.0.0.9000
+## New classes and methods
+* Add `needs()` to check for the availability of a package.
 
 # arkhe 1.0.0
 ## New classes and methods